+++ examples/resources/sdwan_system_feature_profile/resource.tf
@@ -1,0 +1,4 @@
+resource "sdwan_system_feature_profile" "example" {
+  name        = "SYSTEM_FP_1"
+  description = "My system feature profile 1"
+}