+++ docs/guides/changelog.md
@@ -7,6 +7,10 @@ description: |-
 
 # Changelog
 
+## 0.2.12 (unreleased)
+
+- Add `sdwan_system_feature_profile` resource and data source
+
 ## 0.2.11
 
 - Fix idempotency issue with various resource and boolean values